Basic packet manager - sender, receiver, waiter (#7)
* Draft of packet_manager

* Draft of tests for receiver

* More tests of receiver

* Added tests for sender

* Added basic test for writer and temporarily switched channels to use structs insead of pointers

* Renamed Writer to Waiter.
Added Request channel to receiver.
Added more test.
Minor fixes.

+++ src/networking/peer_conn/peer_conn_test.go
@@ -174,7 +174,7 @@ func TestEncodeError(t *testing.T) {
 	length := m[5:7]
 	err := m[7 : 7+utility.GetNumberFromBytes(length)]
 
-	if utility.GetMessageType(m) != uint16(typeMessage[0]) {
+	if utility.GetMessageType(m) != uint8(typeMessage[0]) {
 		t.Errorf(
 			"Incorrect MessageType, got %d, but expected %d", utility.GetMessageType(m), typeMessage[0])
 		return
